Added middleware error handeling for the nonprofit routes

+++ backend/routes/nonprofit.js
@@ -4,6 +4,10 @@ import {
   checkNonProfitName,
   checkNonProfitId,
 } from "#utils/nonprofit_utils.js";
+import {
+  NonProfitNotFoundError,
+  NonProfitAlreadyExistsError,
+} from "#errors/nonprofit-errors.js";
 const prisma = new PrismaClient();
 const nonprofitRouter = express.Router();
 
@@ -13,23 +17,50 @@ nonprofitRouter.get("/", async (req, res) => {
 });
 
 // Get all nonprofits
-nonprofitRouter.get("/all", async (req, res) => {
-  const foundNonProfits = await prisma.nonprofit.findMany();
-  res.status(200).json(foundNonProfits);
+nonprofitRouter.get("/all", async (req, res, next) => {
+  try {
+    const foundNonProfits = await prisma.nonprofit.findMany();
+    res.status(200).json(foundNonProfits);
+  } catch (e) {
+    return next(e);
+  }
 });
 
 // Get one nonprofit by name
-nonprofitRouter.get("/:nonprofit_id", async (req, res) => {
+nonprofitRouter.get("/:nonprofit_name", async (req, res, next) => {
+  const { nonprofit_name } = req.params;
+  try {
+    const findNonProfit = await prisma.nonprofit.findUnique({
+      where: {
+        name: nonprofit_name,
+      },
+    });
+    if (findNonProfit) {
+      res.status(200).json(findNonProfit);
+    } else {
+      throw new NonProfitNotFoundError(nonprofit_name);
+    }
+  } catch (e) {
+    return next(e);
+  }
+});
+
+// Get one nonprofit by id
+nonprofitRouter.get("/id/:nonprofit_id", async (req, res) => {
   const { nonprofit_id } = req.params;
-  const findNonProfit = await prisma.nonprofit.findUnique({
-    where: {
-      id: nonprofit_id,
-    },
-  });
-  if (findNonProfit) {
-    res.status(200).json(findNonProfit);
-  } else {
-    res.status(404).send(`Nonprofit ${nonprofit_id} not found`);
+  try {
+    const findNonProfit = await prisma.nonprofit.findUnique({
+      where: {
+        id: nonprofit_id,
+      },
+    });
+    if (findNonProfit) {
+      res.status(200).json(findNonProfit);
+    } else {
+      throw new NonProfitNotFoundError(nonprofit_id);
+    }
+  } catch (e) {
+    return next(e);
   }
 });
 
@@ -37,15 +68,19 @@ nonprofitRouter.get("/:nonprofit_id", async (req, res) => {
 nonprofitRouter.post("/add", async (req, res) => {
   const nonProfitData = req.body;
   const name = nonProfitData.name;
-  const exists = await checkNonProfitName(name);
+  try {
+    const exists = await checkNonProfitName(name, next);
 
-  if (!exists) {
-    const createNonProfit = await prisma.nonprofit.create({
-      data: nonProfitData,
-    });
-    res.status(201).json(createNonProfit);
-  } else {
-    res.status(400).send(`Nonprofit ${name} already exists`);
+    if (!exists) {
+      const createNonProfit = await prisma.nonprofit.create({
+        data: nonProfitData,
+      });
+      res.status(201).json(createNonProfit);
+    } else {
+      throw new NonProfitAlreadyExistsError(name);
+    }
+  } catch (e) {
+    return next(e);
   }
 });
 
@@ -54,31 +89,54 @@ nonprofitRouter.put("/:nonprofitname/edit", async (req, res) => {
   const { nonprofitname } = req.params;
   const nonProfitData = req.body;
   try {
-    const updateOne = await prisma.nonprofit.update({
-      where: {
-        id: nonprofit_id,
-      },
-      data: nonProfitData,
-    });
-    res.json(updateOne);
-    res.status(214).send();
-  } catch (PrismaClientKnownRequestError) {
-    res.status(404).send("Nonprofit not found");
+    const exists = await checkNonProfitId(nonprofit_id, next);
+
+    if (exists) {
+      const updateOne = await prisma.nonprofit.update({
+        where: {
+          id: nonprofit_id,
+        },
+        data: nonProfitData,
+      });
+      res.json(updateOne);
+      res.status(214).send();
+    } else {
+      throw NonProfitNotFoundError(nonprofit_id);
+    }
+  } catch (e) {
+    return next(e);
   }
 });
 
-// Delete a nonprofit by name
-nonprofitRouter.delete("/:nonprofitname/delete", async (req, res) => {
-  const { nonprofitname } = req.params;
+// Delete a nonprofit by id
+nonprofitRouter.delete("/:nonprofit_id/delete", async (req, res) => {
+  const { nonprofit_id } = req.params;
   try {
-    await prisma.nonprofit.delete({
-      where: {
-        id: nonprofit_id,
-      },
-    });
-    res.status(214).send();
-  } catch (PrismaClientKnownRequestError) {
-    res.status(404).send("Nonprofit not found");
+    const exists = await checkNonProfitId(nonprofit_id, next);
+
+    if (exists) {
+      await prisma.nonprofit.delete({
+        where: {
+          id: nonprofit_id,
+        },
+      });
+      res.status(214).send(`Nonprofit ${nonprofit_id} deleted`);
+    } else {
+      throw NonProfitNotFoundError(nonprofit_id);
+    }
+  } catch (e) {
+    return next(e);
+  }
+});
+
+nonprofitRouter.use((err, req, res, next) => {
+  const retStr = `${err.name}: ${err.message}`;
+  if (err instanceof NonProfitNotFoundError) {
+    return res.status(404).send(retStr);
+  } else if (err instanceof NonProfitAlreadyExistsError) {
+    return res.status(409).send(retStr);
+  } else {
+    return res.status(500).send(retStr);
   }
 });
